requirements.txt
appdirs==1.4.4
attrs==21.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
Bottleneck @ file:///C:/ci/bottleneck_1607557040328/work
cassandra-driver==3.25.0
certifi==2020.6.20
chardet==3.0.4
click @ file:///tmp/build/80754af9/click_1621604852318/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
distlib==0.3.1
filelock==3.0.12
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
Flask-Cors==3.0.10
geomet==0.2.1.post1
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.1
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
ipython @ file:///C:/ci/ipython_1628244015142/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
jedi @ file:///C:/ci/jedi_1611341083684/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1624781299557/work
joblib==1.0.1
kiwisolver==1.3.0
MarkupSafe @ file:///C:/ci/markupsafe_1621528502553/work
matplotlib @ file:///C:/Users/Mouli/Desktop/python_work/matplotlib-3.3.2-cp39-cp39-win_amd64.whl
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
missingno==0.5.0
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
multimethod==1.4
networkx==2.6.2
numexpr @ file:///C:/ci/numexpr_1618856728739/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271900803/work
pandas @ file:///C:/ci/pandas_1627570332296/work
pandas-profiling==3.0.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
phik==0.12.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.0.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
pydantic==1.8.2
pygal==2.4.0
pygal-maps-world==1.0.2
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing==2.4.7
pyshp==2.1.2
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.0
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.1
tangled-up-in-unicode==0.1.0
threadpoolctl==2.2.0
tqdm==4.62.2
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.10.0.0
urllib3==1.26.2
virtualenv==20.1.0
virtualenvwrapper-win==1.2.6
visions==0.7.1
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1625570634446/work
